+++ src/p_saveg.c
@@ -22,6 +22,7 @@
 //-----------------------------------------------------------------------------
 
 
+#include <stdint.h>
 #include "doom_config.h"
 
 #include "i_system.h"
@@ -33,7 +34,7 @@
 
 // Pads save_p to a 4-byte boundary
 //  so that the load/save works on SGI&Gecko.
-#define PADSAVEP() save_p += (4 - ((long long)save_p & 3)) & 3
+#define PADSAVEP() save_p += (4 - ((uintptr_t)save_p & 3)) & 3
 
 
 byte* save_p;
@@ -99,7 +100,7 @@ void P_UnArchivePlayers(void)
             if (players[i].psprites[j].state)
             {
                 players[i].psprites[j].state
-                    = &states[(long long)players[i].psprites[j].state];
+                    = &states[(uintptr_t)players[i].psprites[j].state];
             }
         }
     }
@@ -291,11 +292,11 @@ void P_UnArchiveThinkers(void)
                 mobj = Z_Malloc(sizeof(*mobj), PU_LEVEL, 0);
                 doom_memcpy(mobj, save_p, sizeof(*mobj));
                 save_p += sizeof(*mobj);
-                mobj->state = &states[(long long)mobj->state];
+                mobj->state = &states[(uintptr_t)mobj->state];
                 mobj->target = 0;
                 if (mobj->player)
                 {
-                    mobj->player = &players[(long long)mobj->player - 1];
+                    mobj->player = &players[(uintptr_t)mobj->player - 1];
                     mobj->player->mo = mobj;
                 }
                 P_SetThingPosition(mobj);
@@ -492,7 +493,7 @@ void P_UnArchiveSpecials(void)
                 ceiling = Z_Malloc(sizeof(*ceiling), PU_LEVEL, 0);
                 doom_memcpy(ceiling, save_p, sizeof(*ceiling));
                 save_p += sizeof(*ceiling);
-                ceiling->sector = &sectors[(long long)ceiling->sector];
+                ceiling->sector = &sectors[(uintptr_t)ceiling->sector];
                 ceiling->sector->specialdata = ceiling;
 
                 if (ceiling->thinker.function.acp1)
@@ -507,7 +508,7 @@ void P_UnArchiveSpecials(void)
                 door = Z_Malloc(sizeof(*door), PU_LEVEL, 0);
                 doom_memcpy(door, save_p, sizeof(*door));
                 save_p += sizeof(*door);
-                door->sector = &sectors[(long long)door->sector];
+                door->sector = &sectors[(uintptr_t)door->sector];
                 door->sector->specialdata = door;
                 door->thinker.function.acp1 = (actionf_p1)T_VerticalDoor;
                 P_AddThinker(&door->thinker);
@@ -518,7 +519,7 @@ void P_UnArchiveSpecials(void)
                 floor = Z_Malloc(sizeof(*floor), PU_LEVEL, 0);
                 doom_memcpy(floor, save_p, sizeof(*floor));
                 save_p += sizeof(*floor);
-                floor->sector = &sectors[(long long)floor->sector];
+                floor->sector = &sectors[(uintptr_t)floor->sector];
                 floor->sector->specialdata = floor;
                 floor->thinker.function.acp1 = (actionf_p1)T_MoveFloor;
                 P_AddThinker(&floor->thinker);
@@ -529,7 +530,7 @@ void P_UnArchiveSpecials(void)
                 plat = Z_Malloc(sizeof(*plat), PU_LEVEL, 0);
                 doom_memcpy(plat, save_p, sizeof(*plat));
                 save_p += sizeof(*plat);
-                plat->sector = &sectors[(long long)plat->sector];
+                plat->sector = &sectors[(uintptr_t)plat->sector];
                 plat->sector->specialdata = plat;
 
                 if (plat->thinker.function.acp1)
@@ -544,7 +545,7 @@ void P_UnArchiveSpecials(void)
                 flash = Z_Malloc(sizeof(*flash), PU_LEVEL, 0);
                 doom_memcpy(flash, save_p, sizeof(*flash));
                 save_p += sizeof(*flash);
-                flash->sector = &sectors[(long long)flash->sector];
+                flash->sector = &sectors[(uintptr_t)flash->sector];
                 flash->thinker.function.acp1 = (actionf_p1)T_LightFlash;
                 P_AddThinker(&flash->thinker);
                 break;
@@ -554,7 +555,7 @@ void P_UnArchiveSpecials(void)
                 strobe = Z_Malloc(sizeof(*strobe), PU_LEVEL, 0);
                 doom_memcpy(strobe, save_p, sizeof(*strobe));
                 save_p += sizeof(*strobe);
-                strobe->sector = &sectors[(long long)strobe->sector];
+                strobe->sector = &sectors[(uintptr_t)strobe->sector];
                 strobe->thinker.function.acp1 = (actionf_p1)T_StrobeFlash;
                 P_AddThinker(&strobe->thinker);
                 break;
@@ -564,7 +565,7 @@ void P_UnArchiveSpecials(void)
                 glow = Z_Malloc(sizeof(*glow), PU_LEVEL, 0);
                 doom_memcpy(glow, save_p, sizeof(*glow));
                 save_p += sizeof(*glow);
-                glow->sector = &sectors[(long long)glow->sector];
+                glow->sector = &sectors[(uintptr_t)glow->sector];
                 glow->thinker.function.acp1 = (actionf_p1)T_Glow;
                 P_AddThinker(&glow->thinker);
                 break;
