+++ src/r_data.c
@@ -24,6 +24,7 @@
 //-----------------------------------------------------------------------------
 
 
+#include <stdint.h>
 #include "doom_config.h"
 
 #include "i_system.h"
@@ -630,7 +631,7 @@ void R_InitColormaps(void)
     lump = W_GetNumForName("COLORMAP");
     length = W_LumpLength(lump) + 255;
     colormaps = Z_Malloc(length, PU_STATIC, 0);
-    colormaps = (byte*)(((unsigned long long)colormaps + 255) & ~0xff);
+    colormaps = (byte*)(((uintptr_t)colormaps + 255) & ~0xff);
     W_ReadLump(lump, colormaps);
 }
